[REFACTOR] 디렉토리 분리 - 어르신 공고 (recruit)
