+++ src/main/java/com/balybus/galaxy/patient/service/PatientService.java
@@ -12,11 +12,11 @@ public interface PatientService {
 
 
 
-    PatientResponseDto.RecruitHelper recruitHelper(UserDetails userDetails, PatientRequestDto.RecruitHelper dto); // 어르신 공고 등록
-    PatientResponseDto.GetRecruitList getRecruitList(String userEmail,PatientRequestDto.GetRecruitList dto); // 어르신 공고 리스트 조회
-    PatientResponseDto.GetRecruitList getRecruitPersonalList(String userEmail,PatientRequestDto.GetRecruitPersonalList dto); // 어르신 개인 공고 리스트 조회
-    PatientResponseDto.GetOneRecruitPatientInfo getOneRecruitPatientInfo(String userEmail, Long patientLogSeq); // 어르신 정보 상세 조회
-    PatientResponseDto.UpdateRecruitPatientInfo updateRecruitPatientInfo(String userEmail, PatientRequestDto.UpdateRecruitPatientInfo dto); // 어르신 정보 상세 조회
+//    PatientResponseDto.RecruitHelper recruitHelper(UserDetails userDetails, PatientRequestDto.RecruitHelper dto); // 어르신 공고 등록
+//    PatientResponseDto.GetRecruitList getRecruitList(String userEmail,PatientRequestDto.GetRecruitList dto); // 어르신 공고 리스트 조회
+//    PatientResponseDto.GetRecruitList getRecruitPersonalList(String userEmail,PatientRequestDto.GetRecruitPersonalList dto); // 어르신 개인 공고 리스트 조회
+//    PatientResponseDto.GetOneRecruitPatientInfo getOneRecruitPatientInfo(String userEmail, Long patientLogSeq); // 어르신 정보 상세 조회
+//    PatientResponseDto.UpdateRecruitPatientInfo updateRecruitPatientInfo(String userEmail, PatientRequestDto.UpdateRecruitPatientInfo dto); // 어르신 정보 상세 조회
